Move PhysAddr and PhysRange to port
