+++ port/src/lib.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::upper_case_acronyms)]
 #![cfg_attr(not(any(test)), no_std)]
 #![feature(maybe_uninit_slice)]
+#![feature(step_trait)]
 #![forbid(unsafe_op_in_unsafe_fn)]
 
 pub mod dat;
